+++ packages/next/src/server/app-render/rsc/postpone.ts
@@ -5,21 +5,4 @@ Files in the rsc directory are meant to be packaged as part of the RSC graph usi
 */
 
 // When postpone is available in canary React we can switch to importing it directly
-import React from 'react'
-
-const missingPostpone = typeof React.unstable_postpone !== 'function'
-
-/**
- * This component will call `React.postpone` that throws the postponed error.
- */
-type PostponeProps = {
-  reason: string
-}
-export function Postpone({ reason }: PostponeProps): never {
-  if (missingPostpone) {
-    throw new Error(
-      'Invariant: Postpone component expected `postpone` to exist in React.'
-    )
-  }
-  return React.unstable_postpone(reason)
-}
+export { Postpone } from '../dynamic-rendering'
